requirements.txt
jsonschema==4.25.1
jsondiff==2.2.1
pre-commit==4.5.1
pytest==9.0.2
PyYAML==6.0.3
yamllint==1.37.1
gitpython==3.1.45
psutil==7.2.1
